+++ CHANGELOG.md
@@ -7,6 +7,12 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+## [1.9.0] - 2026-03-02
+
+### Added
+
+- Add support for short format method in RepresentableAsString annotation [#412](https://github.com/nbbrd/java-design-util/issues/412)
+
 ## [1.8.0] - 2025-07-25
 
 ### Added
@@ -125,7 +131,8 @@ This release replaces the `StringValue` annotation with a more general approach 
 
 - Initial release
 
-[Unreleased]: https://github.com/nbbrd/java-design-util/compare/v1.8.0...HEAD
+[Unreleased]: https://github.com/nbbrd/java-design-util/compare/v1.9.0...HEAD
+[1.9.0]: https://github.com/nbbrd/java-design-util/compare/v1.8.0...v1.9.0
 [1.8.0]: https://github.com/nbbrd/java-design-util/compare/v1.7.0...v1.8.0
 [1.7.0]: https://github.com/nbbrd/java-design-util/compare/v1.6.0...v1.7.0
 [1.6.0]: https://github.com/nbbrd/java-design-util/compare/v1.5.1...v1.6.0